Move cluster update to a separate file
